requirements.txt
fluidsynth==0.2
ipython==5.7.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.2.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numpy==1.15.2
pandas==0.23.4
pretty-midi==0.2.8
pyFluidSynth==1.2.5
pypianoroll==0.5.0
python-dateutil==2.7.3
scikit-learn==0.20.0
scipy==1.1.0
tensorboard==1.11.0
tensorboardX==1.4
tensorflow-gpu==1.11.0
